+++ src/core/farcasterTypes.ts
@@ -36,3 +36,15 @@ export function convertToFrame(json: any) {
     },
   };
 }
+
+export type FrameButtonMetadata = {
+  label: string;
+  action?: 'post' | 'post_redirect';
+};
+
+export type FrameMetadata = {
+  image: string;
+  buttons?: [FrameButtonMetadata, ...FrameButtonMetadata[]];
+  post_url?: string;
+  refresh_period?: number;
+};