+++ CHANGELOG.md
@@ -1,5 +1,97 @@
 # Changelog
 
+## 0.3.0
+
+### Minor Changes
+
+- **feat** have `getFrameAccountAddress` reading from the message instead of the body. By @zizzamia #46 0695eb9
+
+- **feat** update `getFrameMetadata` to the latest [Frame APIs](https://warpcast.com/v/0x24295a0a) By @zizzamia #43
+
+BREAKING CHANGES
+
+**getFrameAccountAddress**
+We have enhanced the `getFrameAccountAddress` method by making it more composable with `getFrameMessage`. Now, instead of directly retrieving the `accountAddress` from the `body`, you will utilize the validated `message` to do so.
+
+Before
+
+```ts
+import { getFrameAccountAddress } from '@coinbase/onchainkit';
+
+...
+
+const accountAddress = await getFrameAccountAddress(body);
+```
+
+After
+
+```ts
+import { getFrameAccountAddress } from '@coinbase/onchainkit';
+
+...
+const { isValid, message } = await getFrameMessage(body);
+const accountAddress = await getFrameAccountAddress(message);
+```
+
+**getFrameMetadata**
+We have improved the `getFrameMetadata` method by making the `buttons` extensible for new actions.
+
+Before
+
+```ts
+import { getFrameMetadata } from '@coinbase/onchainkit';
+
+...
+const frameMetadata = getFrameMetadata({
+  buttons: ['boat'],
+  image: 'https://build-onchain-apps.vercel.app/release/v-0-17.png',
+  post_url: 'https://build-onchain-apps.vercel.app/api/frame',
+});
+```
+
+```ts
+type FrameMetadata = {
+  buttons: string[];
+  image: string;
+  post_url: string;
+};
+```
+
+After
+
+```ts
+import { frameMetadata } from '@coinbase/onchainkit';
+
+...
+const frameMetadata = getFrameMetadata({
+  buttons: [
+    {
+      label: 'We love BOAT',
+    },
+  ],
+  image: 'https://build-onchain-apps.vercel.app/release/v-0-17.png',
+  post_url: 'https://build-onchain-apps.vercel.app/api/frame',
+});
+```
+
+```ts
+type Button = {
+  label: string;
+  action?: 'post' | 'post_redirect';
+};
+
+type FrameMetadata = {
+  // A list of strings which are the label for the buttons in the frame (max 4 buttons).
+  buttons: [Button, ...Button[]];
+  // An image which must be smaller than 10MB and should have an aspect ratio of 1.91:1
+  image: string;
+  // A valid POST URL to send the Signature Packet to.
+  post_url: string;
+  // A period in seconds at which the app should expect the image to update.
+  refresh_period?: number;
+};
+```
+
 ## 0.2.1
 
 ### Patch Changes